\name{chaotic_logistic_series}
\docType{data}
\alias{chaotic_logistic_series}
\title{A List of Time Series}
\description{
  A list of 20 time series constructed by using the Chaotic Logistic Map (May, 1976) data generation process. 
  This is a part of a simulated dataset used in Hewamalage et al.(2021). 
  These series can be used to train the SETAR-Tree and SETAR-Forest models.
}
\usage{chaotic_logistic_series}
\format{A list containing 20 numerical vectors.}
\keyword{datasets}
\references{
May, R. M. (1976). Simple mathematical models with very complicated dynamics. Nature, 261, 459–467.

Hewamalage, H., Bergmeir, C., & Bandara, K. (2021). Global models for time series forecasting: A simulation study. Pattern Recognition, 108441.
}