% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{forecast.setarforest}
\alias{forecast.setarforest}
\title{Forecast method for SETAR-Forest fits}
\usage{
\method{forecast}{setarforest}(object, newdata, h = 5)
}
\arguments{
\item{object}{An object of class 'setarforest' which is a trained SETAR-Forest model.}

\item{newdata}{A list of time series which need forecasts or a dataframe/matrix of new instances which need predictions.}

\item{h}{The required number of forecasts (forecast horizon). This parameter is only required when 'newdata' is a list of time series. Default value is 5.}
}
\value{
If 'newdata' is a list of time series, then a list of objects of class 'forecast' is returned. The 'plot' function in the R 'forecast' package can then be used to produce a plot of any time series in the returned list. Each list object contains the following properties.
\item{method}{The name of the forecasting method (SETAR-Forest) as a character string.}
\item{x}{The original time series.}
\item{mean}{Point forecasts as a time series.}
If 'newdata' is a dataframe/matrix, then a vector containing the prediction of each instance is returned.
}
\description{
Obtains forecasts for a given set of time series or a dataframe/matrix of new instances from a fitted SETAR-Forest model.
}
\examples{
# Obtaining forecasts for a list of time series
forest1 <- setarforest(chaotic_logistic_series, bagging_freq = 2)
forecast(forest1, chaotic_logistic_series)

# Obtaining forecasts for a set of test instances
forest2 <- setarforest(data = web_traffic_train[,-1],
                       label = web_traffic_train[,1],
                       bagging_freq = 2,
                       categorical_covariates = c("Project", "Access", "Agent"))
forecast(forest2, web_traffic_test)

}
