% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{forecast.setartree}
\alias{forecast.setartree}
\title{Forecast method for SETAR-Tree fits}
\usage{
\method{forecast}{setartree}(tree, newdata, h = 5)
}
\arguments{
\item{tree}{An object of class 'setartree' which is a trained SETAR-Tree model.}

\item{newdata}{A list of time series which needs forecasts or a dataframe/matrix of new instances which need predictions.}

\item{h}{The required number of forecasts (forecast horizon). This parameter is only required when 'newdata' is a list of time series. Default value is 5.}
}
\value{
If 'newdata' is a list of time series, then a list of objects of class 'forecast' is returned. The 'plot' function can then be used to produce a plot of any time series in the returned list. Each list object contains the following properties.
\item{method}{The name of the forecasting method (SETAR-Tree) as a character string.}
\item{x}{The original time series.}
\item{mean}{Point forecasts as a time series.}
If 'newdata' is a dataframe/matrix, then a vector containing the prediction of each instance is returned.
}
\description{
Obtains forecasts for a given set of time series or a dataframe/matrix of new instances from a fitted SETAR-Tree model.
}
\examples{
# Obtaining forecasts for a list of time series
tree1 <- setartree(train_series)
forecasts <- forecast(tree1, train_series)

plot(forecasts[[1]]) # Plotting the first time series with its forecasts

# Obtaining forecasts for a set of test instances
tree2 <- setartree(data = train_df[,-1],
                   label = train_df[,1],
                   stopping_criteria = "error_imp",
                   categorical_covariates = c("Open", "Promo", "StateHoliday", "SchoolHoliday"))
forecast(tree2, test_df)

}
