\name{test_df}
\docType{data}
\alias{test_df}
\title{A Dataframe of Test Instances}
\description{
  A dataframe containing 10 instances that can be used to test the SETAR-Tree and SETAR-Forest models. Each intance consists of 10 time series lags (Lag1 to Lag10), 4 categorical covariates (Open, Promo, StateHoliday and SchoolHoliday) and a numerical covariate (Customers).
}
\usage{test_df}
\format{A dataframe containing 10 test instances.}
\keyword{datasets}