\name{train_df}
\docType{data}
\alias{train_df}
\title{A Dataframe of Training Instances}
\description{
  A dataframe containing 100 instances that can be used to train the SETAR-Tree and SETAR-Forest models. Each intance consists of 10 time series lags (Lag1 to Lag10), 4 categorical covariates (Open, Promo, StateHoliday and SchoolHoliday), a numerical covariate (Customers) and the corresponding true outputs (y).
}
\usage{train_df}
\format{A dataframe containing 100 training instances.}
\keyword{datasets}