% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docData.R
\docType{data}
\name{web_traffic_test}
\alias{web_traffic_test}
\title{A dataframe of test instances}
\format{
A dataframe containing 5 test instances.
}
\description{
A dataframe containing 5 instances that can be used to test the SETAR-Tree and SETAR-Forest models.
The data are related to the number of hits or web traffic of a set of Wikipedia pages.
Each intance consists of 10 time series lags (Lag1 to Lag10) and a categorical covariate (Project).
The data were downloaded from the Wikimedia REST API (Wikimedia, 2022).
}
\examples{
web_traffic_test
}
\references{
Wikimedia Analytics Team (2022). Wikistats: Pageview complete dumps.\cr
URL https://dumps.wikimedia.org/other/pageview_complete
}
\keyword{datasets}
